+++ Node/src/RingBuffer.cpp
@@ -2,7 +2,7 @@
 
 RingBuffer::RingBuffer(int size) {
     capacity = size;
-    buffer = new uint8_t[capacity];
+    buffer = new uint16_t[capacity];
     head = 0;
     tail = 0;
     count = 0;
@@ -12,7 +12,7 @@ RingBuffer::~RingBuffer() {
     delete[] buffer;
 }
 
-void RingBuffer::enqueue(uint8_t value) {
+void RingBuffer::enqueue(uint16_t value) {
     buffer[head] = value;
     head = (head + 1) % capacity;
     if (count == capacity) {
@@ -22,22 +22,22 @@ void RingBuffer::enqueue(uint8_t value) {
     }
 }
 
-uint8_t RingBuffer::dequeue() {
+uint16_t RingBuffer::dequeue() {
     if (count == 0) return 255; // Return invalid if empty
-    uint8_t value = buffer[tail];
+    uint16_t value = buffer[tail];
     tail = (tail + 1) % capacity;
     count--;
     return value;
 }
 
-bool RingBuffer::contains(uint8_t value) {
+bool RingBuffer::contains(uint16_t value) {
     for (int i = 0, idx = tail; i < count; i++, idx = (idx + 1) % capacity) {
         if (buffer[idx] == value) return true;
     }
     return false;
 }
 
-bool RingBuffer::remove(uint8_t value) {
+bool RingBuffer::remove(uint16_t value) {
     if (count == 0) return false;
 
     int shiftCount = 0;
